Move Speed type to usb crate and distinguish SplitSpeed.
