+++ src/backend/mod.rs
@@ -9,10 +9,10 @@ use anyhow::{Context, Error};
 use futures_channel::oneshot;
 use futures_lite::future::block_on;
 use nusb::{self, DeviceInfo};
-use num_enum::{FromPrimitive, IntoPrimitive};
 use once_cell::sync::Lazy;
 
 use crate::util::handle_thread_panic;
+pub use crate::usb::Speed;
 
 pub mod cynthion;
 pub mod ice40usbtrace;
@@ -65,30 +65,6 @@ pub trait BackendDevice {
 
     /// Which speeds this device supports.
     fn supported_speeds(&self) -> &[Speed];
-}
-
-/// Possible capture speed settings.
-#[derive(Debug, Copy, Clone, PartialEq, FromPrimitive, IntoPrimitive)]
-#[repr(u8)]
-pub enum Speed {
-    #[default]
-    High = 0,
-    Full = 1,
-    Low  = 2,
-    Auto = 3,
-}
-
-impl Speed {
-    /// How this speed setting should be displayed in the UI.
-    pub fn description(&self) -> &'static str {
-        use Speed::*;
-        match self {
-            Auto => "Auto",
-            High => "High (480Mbps)",
-            Full => "Full (12Mbps)",
-            Low => "Low (1.5Mbps)",
-        }
-    }
 }
 
 /// A timestamped packet.
