+++ src/usb.rs
@@ -49,6 +49,16 @@ pub fn crc5(mut input: u32, num_bits: u32) -> u8 {
     (state ^ 0x1f) as u8
 }
 
+#[derive(Copy, Clone, Debug, FromPrimitive, IntoPrimitive, PartialEq)]
+#[repr(u8)]
+pub enum Speed {
+    #[default]
+    High = 0,
+    Full = 1,
+    Low  = 2,
+    Auto = 3,
+}
+
 #[allow(clippy::upper_case_acronyms)]
 #[derive(Copy, Clone, Debug, Default, IntoPrimitive, FromPrimitive, PartialEq, Eq)]
 #[repr(u8)]
@@ -270,7 +280,7 @@ pub enum StartComplete {
 }
 
 #[derive(Copy, Clone, Debug)]
-pub enum Speed {
+pub enum SplitSpeed {
     Low,
     Full,
 }
@@ -314,8 +324,8 @@ impl SplitFields {
                 [packet[1], packet[2], packet[3], 0]))
     }
 
-    pub fn speed(&self) -> Speed {
-        use Speed::*;
+    pub fn speed(&self) -> SplitSpeed {
+        use SplitSpeed::*;
         if self.endpoint_type() == EndpointType::Isochronous {
             Full
         } else if self.start() {
